refactored ExecutePrimitiveContextNode to be a peer, not a subclass, of ExecuteContextNode
