+++ src/de.hpi.swa.graal.squeak/src/de/hpi/swa/graal/squeak/nodes/ExecuteContextNode.java
@@ -6,128 +6,25 @@
 package de.hpi.swa.graal.squeak.nodes;
 
 import com.oracle.truffle.api.CompilerAsserts;
-import com.oracle.truffle.api.CompilerDirectives;
-import com.oracle.truffle.api.TruffleLogger;
-import com.oracle.truffle.api.dsl.Cached;
-import com.oracle.truffle.api.dsl.Fallback;
-import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.instrumentation.GenerateWrapper;
 import com.oracle.truffle.api.instrumentation.InstrumentableNode;
 import com.oracle.truffle.api.instrumentation.ProbeNode;
 import com.oracle.truffle.api.instrumentation.StandardTags;
 import com.oracle.truffle.api.instrumentation.Tag;
-import com.oracle.truffle.api.nodes.ExplodeLoop;
-import com.oracle.truffle.api.nodes.LoopNode;
-import com.oracle.truffle.api.nodes.NodeCost;
-import com.oracle.truffle.api.nodes.NodeInfo;
-import com.oracle.truffle.api.source.Source;
-import com.oracle.truffle.api.source.SourceSection;
 
-import de.hpi.swa.graal.squeak.exceptions.PrimitiveExceptions.PrimitiveFailed;
-import de.hpi.swa.graal.squeak.exceptions.ProcessSwitch;
-import de.hpi.swa.graal.squeak.exceptions.Returns.NonLocalReturn;
-import de.hpi.swa.graal.squeak.exceptions.Returns.NonVirtualReturn;
 import de.hpi.swa.graal.squeak.model.CompiledCodeObject;
-import de.hpi.swa.graal.squeak.model.CompiledMethodObject;
-import de.hpi.swa.graal.squeak.nodes.ExecuteContextNodeFactory.TriggerInterruptHandlerNodeGen;
-import de.hpi.swa.graal.squeak.nodes.bytecodes.AbstractBytecodeNode;
-import de.hpi.swa.graal.squeak.nodes.bytecodes.JumpBytecodes.ConditionalJumpNode;
-import de.hpi.swa.graal.squeak.nodes.bytecodes.JumpBytecodes.UnconditionalJumpNode;
-import de.hpi.swa.graal.squeak.nodes.bytecodes.MiscellaneousBytecodes.CallPrimitiveNode;
-import de.hpi.swa.graal.squeak.nodes.bytecodes.PushBytecodes.PushClosureNode;
-import de.hpi.swa.graal.squeak.nodes.bytecodes.ReturnBytecodes.AbstractReturnNode;
-import de.hpi.swa.graal.squeak.nodes.bytecodes.SendBytecodes.AbstractSendNode;
-import de.hpi.swa.graal.squeak.nodes.context.frame.FrameStackInitializationNode;
-import de.hpi.swa.graal.squeak.shared.SqueakLanguageConfig;
-import de.hpi.swa.graal.squeak.util.ArrayUtils;
 import de.hpi.swa.graal.squeak.util.FrameAccess;
-import de.hpi.swa.graal.squeak.util.InterruptHandlerNode;
-import de.hpi.swa.graal.squeak.util.SqueakBytecodeDecoder;
 
 @GenerateWrapper
-public class ExecuteContextNode extends AbstractNodeWithCode implements InstrumentableNode {
-
-    private static final TruffleLogger LOG = TruffleLogger.getLogger(SqueakLanguageConfig.ID, CallPrimitiveNode.class);
-    private static final boolean DECODE_BYTECODE_ON_DEMAND = true;
-    private static final int STACK_DEPTH_LIMIT = 25000;
-    private static final int LOCAL_RETURN_PC = -1;
-
-    @Children private AbstractBytecodeNode[] bytecodeNodes;
-    @Child private HandleNonLocalReturnNode handleNonLocalReturnNode;
-    @Child private GetOrCreateContextNode getOrCreateContextNode;
-
-    @Child private FrameStackInitializationNode frameInitializationNode;
-    @Child private MaterializeContextOnMethodExitNode materializeContextOnMethodExitNode;
-
-    private SourceSection section;
-
-    private static final class ExecutePrimitiveContextNode extends ExecuteContextNode {
-
-        @Child private HandlePrimitiveFailedNode handlePrimitiveFailedNode;
-
-        ExecutePrimitiveContextNode(final CompiledCodeObject code, final boolean resume) {
-            super(code, resume);
-            assert !resume || code.isUnwindMarked() || code instanceof CompiledMethodObject && ((CompiledMethodObject) code).isExceptionHandlerMarked();
-        }
-
-        @Override
-        public Object executeFresh(final VirtualFrame frame) {
-            super.initialize(frame);
-            final CallPrimitiveNode callPrimitiveNode = (CallPrimitiveNode) fetchNextBytecodeNode(0);
-            if (callPrimitiveNode.primitiveNode != null) {
-                try {
-                    return callPrimitiveNode.primitiveNode.executePrimitive(frame);
-                } catch (final PrimitiveFailed e) {
-                    getHandlePrimitiveFailedNode().executeHandle(frame, e.getReasonCode());
-                    LOG.finer(() -> callPrimitiveNode.primitiveNode +
-                                    " (" + ArrayUtils.toJoinedString(", ", FrameAccess.getReceiverAndArguments(frame)) + ")");
-                    /* continue with fallback code. */
-                }
-            }
-            assert callPrimitiveNode.getSuccessorIndex() == CallPrimitiveNode.NUM_BYTECODES;
-            return super.executeFresh(frame);
-        }
-
-        @Override
-        protected void initialize(final VirtualFrame frame) {
-            // avoid re-initializing in fallback code
-        }
-
-        @Override
-        protected int startBytecodeIndex() {
-            return CallPrimitiveNode.NUM_BYTECODES;
-        }
-
-        private HandlePrimitiveFailedNode getHandlePrimitiveFailedNode() {
-            if (handlePrimitiveFailedNode == null) {
-                CompilerDirectives.transferToInterpreterAndInvalidate();
-                handlePrimitiveFailedNode = insert(HandlePrimitiveFailedNode.create(code));
-            }
-            return handlePrimitiveFailedNode;
-        }
-    }
+public class ExecuteContextNode extends AbstractExecuteContextNode implements InstrumentableNode {
 
     protected ExecuteContextNode(final CompiledCodeObject code, final boolean resume) {
-        super(code);
-        if (DECODE_BYTECODE_ON_DEMAND) {
-            bytecodeNodes = new AbstractBytecodeNode[SqueakBytecodeDecoder.trailerPosition(code)];
-        } else {
-            bytecodeNodes = SqueakBytecodeDecoder.decode(code);
-        }
-        frameInitializationNode = resume ? null : FrameStackInitializationNode.create(code);
-        materializeContextOnMethodExitNode = resume ? null : MaterializeContextOnMethodExitNode.create(code);
+        super(code, resume);
     }
 
     protected ExecuteContextNode(final ExecuteContextNode executeContextNode) {
-        this(executeContextNode.code, executeContextNode.frameInitializationNode == null);
-    }
-
-    public static ExecuteContextNode create(final CompiledCodeObject code, final boolean resume) {
-        if (code.hasPrimitive()) {
-            return new ExecutePrimitiveContextNode(code, resume);
-        }
-        return new ExecuteContextNode(code, resume);
+        super(executeContextNode);
     }
 
     @Override
@@ -136,187 +33,11 @@ public class ExecuteContextNode extends AbstractNodeWithCode implements Instrume
         return code.toString();
     }
 
-    public Object executeFresh(final VirtualFrame frame) {
-        final int pc = startBytecodeIndex();
-        FrameAccess.setInstructionPointer(frame, code, pc);
-        final boolean enableStackDepthProtection = enableStackDepthProtection();
-        try {
-            if (enableStackDepthProtection && code.image.stackDepth++ > STACK_DEPTH_LIMIT) {
-                throw ProcessSwitch.createWithBoundary(getGetOrCreateContextNode().executeGet(frame));
-            }
-            initialize(frame);
-            return startBytecode(frame, pc);
-        } catch (final NonLocalReturn nlr) {
-            /** {@link getHandleNonLocalReturnNode()} acts as {@link BranchProfile} */
-            return getHandleNonLocalReturnNode().executeHandle(frame, nlr);
-        } catch (final NonVirtualReturn nvr) {
-            /** {@link getGetOrCreateContextNode()} acts as {@link BranchProfile} */
-            getGetOrCreateContextNode().executeGet(frame).markEscaped();
-            throw nvr;
-        } catch (final ProcessSwitch ps) {
-            /** {@link getGetOrCreateContextNode()} acts as {@link BranchProfile} */
-            getGetOrCreateContextNode().executeGet(frame).markEscaped();
-            throw ps;
-        } finally {
-            if (enableStackDepthProtection) {
-                code.image.stackDepth--;
-            }
-            materializeContextOnMethodExitNode.execute(frame);
-        }
-    }
-
-    protected void initialize(final VirtualFrame frame) {
-        frameInitializationNode.executeInitialize(frame);
-    }
-
-    public final Object executeResumeAtStart(final VirtualFrame frame) {
-        assert startBytecodeIndex() == 0;
-        try {
-            return startBytecode(frame, 0);
-        } catch (final NonLocalReturn nlr) {
-            /** {@link getHandleNonLocalReturnNode()} acts as {@link BranchProfile} */
-            return getHandleNonLocalReturnNode().executeHandle(frame, nlr);
-        } finally {
-            code.image.lastSeenContext = null; // Stop materialization here.
-        }
-    }
-
-    protected int startBytecodeIndex() {
-        return 0;
-    }
-
-    public final Object executeResumeInMiddle(final VirtualFrame frame, final long initialPC) {
-        try {
-            return resumeBytecode(frame, initialPC);
-        } catch (final NonLocalReturn nlr) {
-            /** {@link getHandleNonLocalReturnNode()} acts as {@link BranchProfile} */
-            return getHandleNonLocalReturnNode().executeHandle(frame, nlr);
-        } finally {
-            code.image.lastSeenContext = null; // Stop materialization here.
-        }
-    }
-
-    private GetOrCreateContextNode getGetOrCreateContextNode() {
-        if (getOrCreateContextNode == null) {
-            CompilerDirectives.transferToInterpreterAndInvalidate();
-            getOrCreateContextNode = insert(GetOrCreateContextNode.create(code));
-        }
-        return getOrCreateContextNode;
-    }
-
-    /*
-     * Inspired by Sulong's LLVMDispatchBasicBlockNode (https://git.io/fjEDw).
-     */
-    @ExplodeLoop(kind = ExplodeLoop.LoopExplosionKind.MERGE_EXPLODE)
-    private Object startBytecode(final VirtualFrame frame, final int initialPC) {
-        CompilerAsserts.compilationConstant(bytecodeNodes.length);
-        int pc = initialPC;
-        int backJumpCounter = 0;
-        Object returnValue = null;
-        bytecode_loop: while (pc != LOCAL_RETURN_PC) {
-            CompilerAsserts.partialEvaluationConstant(pc);
-            final AbstractBytecodeNode node = fetchNextBytecodeNode(pc);
-            if (node instanceof ConditionalJumpNode) {
-                final ConditionalJumpNode jumpNode = (ConditionalJumpNode) node;
-                final int successor = jumpNode.executeCondition(frame) ? jumpNode.getJumpSuccessorIndex() : jumpNode.getSuccessorIndex();
-                if (CompilerDirectives.inInterpreter() && successor <= pc) {
-                    backJumpCounter++;
-                }
-                pc = successor;
-                continue bytecode_loop;
-            } else if (node instanceof AbstractReturnNode) {
-                returnValue = ((AbstractReturnNode) node).executeReturn(frame);
-                pc = LOCAL_RETURN_PC;
-                continue bytecode_loop;
-            } else if (node instanceof UnconditionalJumpNode) {
-                final int successor = ((UnconditionalJumpNode) node).getJumpSuccessor();
-                if (CompilerDirectives.inInterpreter() && successor <= pc) {
-                    backJumpCounter++;
-                }
-                pc = successor;
-                continue bytecode_loop;
-            } else if (node instanceof PushClosureNode) {
-                final PushClosureNode pushClosureNode = (PushClosureNode) node;
-                pushClosureNode.executePush(frame);
-                pc = pushClosureNode.getClosureSuccessorIndex();
-                continue bytecode_loop;
-            } else {
-                pc = node.getSuccessorIndex();
-                if (node instanceof AbstractSendNode) {
-                    FrameAccess.setInstructionPointer(frame, code, pc);
-                }
-                // TODO - for Sista, this would be a good place to deal with inlined primitives
-                node.executeVoid(frame);
-                continue bytecode_loop;
-            }
-        }
-        assert returnValue != null && !hasModifiedSender(frame);
-        FrameAccess.terminate(frame, code);
-        assert backJumpCounter >= 0;
-        LoopNode.reportLoopCount(this, backJumpCounter);
-        return returnValue;
-    }
-
-    /*
-     * Non-optimized version of startBytecode which is used to resume contexts.
-     */
-    private Object resumeBytecode(final VirtualFrame frame, final long initialPC) {
-        assert initialPC > 0 : "Trying to resume a fresh/terminated/illegal context";
-        int pc = (int) initialPC;
-        Object returnValue = null;
-        bytecode_loop_slow: while (pc != LOCAL_RETURN_PC) {
-            final AbstractBytecodeNode node = fetchNextBytecodeNode(pc);
-            if (node instanceof ConditionalJumpNode) {
-                final ConditionalJumpNode jumpNode = (ConditionalJumpNode) node;
-                pc = jumpNode.executeCondition(frame) ? jumpNode.getJumpSuccessorIndex() : jumpNode.getSuccessorIndex();
-                continue bytecode_loop_slow;
-            } else if (node instanceof AbstractReturnNode) {
-                returnValue = ((AbstractReturnNode) node).executeReturn(frame);
-                pc = LOCAL_RETURN_PC;
-                continue bytecode_loop_slow;
-            } else if (node instanceof UnconditionalJumpNode) {
-                pc = ((UnconditionalJumpNode) node).getJumpSuccessor();
-                continue bytecode_loop_slow;
-            } else if (node instanceof PushClosureNode) {
-                final PushClosureNode pushClosureNode = (PushClosureNode) node;
-                pushClosureNode.executePush(frame);
-                pc = pushClosureNode.getClosureSuccessorIndex();
-                continue bytecode_loop_slow;
-            } else {
-                pc = node.getSuccessorIndex();
-                if (node instanceof AbstractSendNode) {
-                    FrameAccess.setInstructionPointer(frame, code, pc);
-                }
-                node.executeVoid(frame);
-                continue bytecode_loop_slow;
-            }
-        }
-        assert returnValue != null && !hasModifiedSender(frame);
-        FrameAccess.terminate(frame, code);
-        return returnValue;
-    }
-
-    /*
-     * Fetch next bytecode and insert AST nodes on demand if enabled.
-     */
-    @SuppressWarnings("unused")
-    protected final AbstractBytecodeNode fetchNextBytecodeNode(final int pc) {
-        if (DECODE_BYTECODE_ON_DEMAND && bytecodeNodes[pc] == null) {
-            CompilerDirectives.transferToInterpreterAndInvalidate();
-            bytecodeNodes[pc] = insert(SqueakBytecodeDecoder.decodeBytecode(code, pc));
-            notifyInserted(bytecodeNodes[pc]);
-        }
-        return bytecodeNodes[pc];
-    }
-
-    /* Only use stackDepthProtection in interpreter or once per compilation unit (if at all). */
-    private boolean enableStackDepthProtection() {
-        return code.image.options.enableStackDepthProtection && (CompilerDirectives.inInterpreter() || CompilerDirectives.inCompilationRoot());
-    }
-
     @Override
-    public final boolean isInstrumentable() {
-        return true;
+    public Object executeFresh(final VirtualFrame frame) {
+        FrameAccess.setInstructionPointer(frame, code, 0);
+        initializeFrame(frame);
+        return doExecuteFresh(frame);
     }
 
     @Override
@@ -330,72 +51,13 @@ public class ExecuteContextNode extends AbstractNodeWithCode implements Instrume
     }
 
     @Override
-    public String getDescription() {
-        return code.toString();
+    public final boolean isInstrumentable() {
+        return true;
     }
 
     @Override
-    public SourceSection getSourceSection() {
-        if (section == null) {
-            if (code.image.isTesting()) {
-                // Cannot provide source section in case of AbstractSqueakTestCaseWithDummyImage.
-                return null;
-            }
-            final Source source = code.getSource();
-            if (source.getCharacters().equals(CompiledCodeObject.SOURCE_UNAVAILABLE)) {
-                section = source.createUnavailableSection();
-            } else {
-                section = source.createSection(1, 1, source.getLength());
-            }
-        }
-        return section;
+    protected final int startPc() {
+        return 0;
     }
 
-    // FIXME: Trigger interrupt check on sends and in loops.
-    @NodeInfo(cost = NodeCost.NONE)
-    protected abstract static class TriggerInterruptHandlerNode extends AbstractNodeWithCode {
-        protected static final int BYTECODE_LENGTH_THRESHOLD = 32;
-
-        protected TriggerInterruptHandlerNode(final CompiledCodeObject code) {
-            super(code);
-        }
-
-        @SuppressWarnings("unused")
-        private static TriggerInterruptHandlerNode create(final CompiledCodeObject code) {
-            return TriggerInterruptHandlerNodeGen.create(code);
-        }
-
-        protected abstract void executeGeneric(VirtualFrame frame);
-
-        @SuppressWarnings("unused")
-        @Specialization(guards = {"shouldTrigger(frame)"})
-        protected static final void doTrigger(final VirtualFrame frame, @Cached("create(code)") final InterruptHandlerNode interruptNode) {
-            interruptNode.executeTrigger(frame);
-        }
-
-        @SuppressWarnings("unused")
-        @Fallback
-        protected final void doNothing() {
-            // Do not trigger.
-        }
-
-        protected final boolean shouldTrigger(@SuppressWarnings("unused") final VirtualFrame frame) {
-            if (CompilerDirectives.inCompiledCode() && !CompilerDirectives.inCompilationRoot()) {
-                return false; // never trigger in inlined code
-            }
-            return code.image.interrupt.isActiveAndShouldTrigger();
-        }
-    }
-
-    private HandleNonLocalReturnNode getHandleNonLocalReturnNode() {
-        if (handleNonLocalReturnNode == null) {
-            CompilerDirectives.transferToInterpreterAndInvalidate();
-            handleNonLocalReturnNode = insert(HandleNonLocalReturnNode.create(code));
-        }
-        return handleNonLocalReturnNode;
-    }
-
-    protected final FrameStackInitializationNode getFrameInitializationNode() {
-        return frameInitializationNode;
-    }
 }
